Extract Verifier enums out of Verify
